+++ Proyecto_Api/proyecto_web/src/main/java/com/proyecto/proyecto_web/model/rol.java
@@ -12,15 +12,15 @@ public class rol {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idrol;
+    private Integer idrol;
 
     private String nombre;
 
-    public long getIdrol() {
+    public Integer getIdrol() {
         return idrol;
     }
 
-    public void setIdrol(long idrol) {
+    public void setIdrol(Integer idrol) {
         this.idrol = idrol;
     }
